Add CounterSign1 to code base (#117)
* Add CounterSign0 to code base

* Fix compiler errors

* Fix context compiler issues

* Fix testing errors

* Fix 2 compiler errors and a crash

* See if the crashes go away

* More bug fixes

* It would be nice if both compilers look at this the same

* Update SignerInfo.cpp

stupid gcc compiler.  does not actually do this right

* More memory leak checks

* Fix bug

correct free code

* Deal with counterisign1 issues

* Context needs to be ifdef dependent

* Remove unused function call

* Allocation failures

* Fix errors from new test cases

* Correct building COSE from EC key for OpenSSL

* Update build scripts

Build countersign1 on appveyor and github
Correct the path from the countersign1 rename in cmake.

* Increase coverage

* FIx apveyor build script
* Add countersign1 testing to Sign1

+++ src/cose_crypto.h
@@ -38,7 +38,7 @@ bool AES_KW_Decrypt(COSE_Enveloped* pcose,
 	const byte* pbCipherText,
 	size_t cbCipherText,
 	byte* pbKeyOut,
-	int* pcbKeyOut,
+	size_t* pcbKeyOut,
 	cose_errback* perr);
 
 /**
@@ -175,8 +175,8 @@ bool ECDSA_Verify(COSE* pSigner,
 	cose_errback* perr);
 
 bool ECDH_ComputeSecret(COSE* pReciient,
-	COSE_KEY **ppKeyMe,
-	COSE_KEY *pKeyYou,
+	COSE_KEY** ppKeyMe,
+	COSE_KEY* pKeyYou,
 	byte** ppbSecret,
 	size_t* pcbSecret,
 	CBOR_CONTEXT_COMMA cose_errback* perr);
